# frigappapi/src/main/java/com/frigapp/frigappapi/controller/api/FridgesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.frigapp.frigappapi.controller.api;

import com.frigapp.frigappapi.model.*;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.context.request.NativeWebRequest;

import java.util.Optional;
import java.util.UUID;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-26T22:36:13.511413300+01:00[Europe/Paris]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "Fridges", description = "Création et gestion des frigos")
public interface FridgesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * DELETE /fridges/{fridgeId} : Supprimer un frigo
     * Supprime le frigo et toutes ses données en cascade. Réservé au propriétaire.
     *
     * @param fridgeId Identifiant du frigo (required)
     * @return Frigo supprimé (status code 204)
     *         or Token JWT absent ou invalide (status code 401)
     *         or Action non autorisée pour ce rôle (status code 403)
     *         or Ressource introuvable (status code 404)
     */
    @Operation(
        operationId = "fridgesFridgeIdDelete",
        summary = "Supprimer un frigo",
        description = "Supprime le frigo et toutes ses données en cascade. Réservé au propriétaire.",
        tags = { "Fridges" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Frigo supprimé"),
            @ApiResponse(responseCode = "401", description = "Token JWT absent ou invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "Action non autorisée pour ce rôle", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Ressource introuvable", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/fridges/{fridgeId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Void> fridgesFridgeIdDelete(
        @Parameter(name = "fridgeId", description = "Identifiant du frigo", required = true, in = ParameterIn.PATH) @PathVariable("fridgeId") UUID fridgeId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /fridges/{fridgeId} : Détail d&#39;un frigo
     *
     * @param fridgeId Identifiant du frigo (required)
     * @return Détail du frigo (status code 200)
     *         or Token JWT absent ou invalide (status code 401)
     *         or Action non autorisée pour ce rôle (status code 403)
     *         or Ressource introuvable (status code 404)
     */
    @Operation(
        operationId = "fridgesFridgeIdGet",
        summary = "Détail d'un frigo",
        tags = { "Fridges" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Détail du frigo", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = FridgesPost201Response.class))
            }),
            @ApiResponse(responseCode = "401", description = "Token JWT absent ou invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "Action non autorisée pour ce rôle", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Ressource introuvable", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/fridges/{fridgeId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<FridgesPost201Response> fridgesFridgeIdGet(
        @Parameter(name = "fridgeId", description = "Identifiant du frigo", required = true, in = ParameterIn.PATH) @PathVariable("fridgeId") UUID fridgeId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"role\" : \"owner\", \"urgentCount\" : 2, \"memberCount\" : 3, \"name\" : \"Frigo maison\", \"id\" : \"550e8400-e29b-41d4-a716-446655440001\", \"ownerId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"itemCount\" : 12, \"updatedAt\" : \"2026-03-23T09:00:00Z\" }, \"status\" : \"success\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /fridges/{fridgeId}/leave : Quitter un frigo
     * Retire l&#39;utilisateur courant des membres du frigo. Réservé aux collaborateurs.
     *
     * @param fridgeId  (required)
     * @return Frigo quitté (status code 204)
     *         or Token JWT absent ou invalide (status code 401)
     *         or Le propriétaire ne peut pas quitter son propre frigo (status code 403)
     *         or Ressource introuvable (status code 404)
     */
    @Operation(
        operationId = "fridgesFridgeIdLeaveDelete",
        summary = "Quitter un frigo",
        description = "Retire l'utilisateur courant des membres du frigo. Réservé aux collaborateurs.",
        tags = { "Fridges" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Frigo quitté"),
            @ApiResponse(responseCode = "401", description = "Token JWT absent ou invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "Le propriétaire ne peut pas quitter son propre frigo", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Ressource introuvable", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/fridges/{fridgeId}/leave",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Void> fridgesFridgeIdLeaveDelete(
        @Parameter(name = "fridgeId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("fridgeId") UUID fridgeId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /fridges/{fridgeId} : Renommer un frigo
     * Réservé au propriétaire du frigo.
     *
     * @param fridgeId Identifiant du frigo (required)
     * @param fridgeRequest  (required)
     * @return Frigo renommé (status code 200)
     *         or Données invalides (status code 400)
     *         or Token JWT absent ou invalide (status code 401)
     *         or Action non autorisée pour ce rôle (status code 403)
     *         or Ressource introuvable (status code 404)
     */
    @Operation(
        operationId = "fridgesFridgeIdPut",
        summary = "Renommer un frigo",
        description = "Réservé au propriétaire du frigo.",
        tags = { "Fridges" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Frigo renommé", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = FridgesPost201Response.class))
            }),
            @ApiResponse(responseCode = "400", description = "Données invalides", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Token JWT absent ou invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "Action non autorisée pour ce rôle", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Ressource introuvable", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/fridges/{fridgeId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<FridgesPost201Response> fridgesFridgeIdPut(
        @Parameter(name = "fridgeId", description = "Identifiant du frigo", required = true, in = ParameterIn.PATH) @PathVariable("fridgeId") UUID fridgeId,
        @Parameter(name = "FridgeRequest", description = "", required = true) @Valid @RequestBody FridgeRequest fridgeRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"role\" : \"owner\", \"urgentCount\" : 2, \"memberCount\" : 3, \"name\" : \"Frigo maison\", \"id\" : \"550e8400-e29b-41d4-a716-446655440001\", \"ownerId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"itemCount\" : 12, \"updatedAt\" : \"2026-03-23T09:00:00Z\" }, \"status\" : \"success\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /fridges : Lister mes frigos
     * Retourne tous les frigos auxquels l&#39;utilisateur appartient (possédés + collaborations).
     *
     * @return Liste des frigos (status code 200)
     *         or Token JWT absent ou invalide (status code 401)
     */
    @Operation(
        operationId = "fridgesGet",
        summary = "Lister mes frigos",
        description = "Retourne tous les frigos auxquels l'utilisateur appartient (possédés + collaborations).",
        tags = { "Fridges" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Liste des frigos", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = FridgesGet200Response.class))
            }),
            @ApiResponse(responseCode = "401", description = "Token JWT absent ou invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/fridges",
        produces = { "application/json" }
    )
    
    default ResponseEntity<FridgesGet200Response> fridgesGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : [ { \"role\" : \"owner\", \"urgentCount\" : 2, \"memberCount\" : 3, \"name\" : \"Frigo maison\", \"id\" : \"550e8400-e29b-41d4-a716-446655440001\", \"itemCount\" : 12, \"updatedAt\" : \"2026-03-23T09:00:00Z\" }, { \"role\" : \"owner\", \"urgentCount\" : 2, \"memberCount\" : 3, \"name\" : \"Frigo maison\", \"id\" : \"550e8400-e29b-41d4-a716-446655440001\", \"itemCount\" : 12, \"updatedAt\" : \"2026-03-23T09:00:00Z\" } ], \"status\" : \"success\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /fridges : Créer un frigo
     * Crée un nouveau frigo. L&#39;utilisateur courant devient automatiquement propriétaire.
     *
     * @param fridgeRequest  (required)
     * @return Frigo créé (status code 201)
     *         or Données invalides (status code 400)
     *         or Token JWT absent ou invalide (status code 401)
     */
    @Operation(
        operationId = "fridgesPost",
        summary = "Créer un frigo",
        description = "Crée un nouveau frigo. L'utilisateur courant devient automatiquement propriétaire.",
        tags = { "Fridges" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Frigo créé", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = FridgesPost201Response.class))
            }),
            @ApiResponse(responseCode = "400", description = "Données invalides", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Token JWT absent ou invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/fridges",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<FridgesPost201Response> fridgesPost(
        @Parameter(name = "FridgeRequest", description = "", required = true) @Valid @RequestBody FridgeRequest fridgeRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"role\" : \"owner\", \"urgentCount\" : 2, \"memberCount\" : 3, \"name\" : \"Frigo maison\", \"id\" : \"550e8400-e29b-41d4-a716-446655440001\", \"ownerId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"itemCount\" : 12, \"updatedAt\" : \"2026-03-23T09:00:00Z\" }, \"status\" : \"success\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }
}