# frigappapi/src/main/resources/openapi/openapi_frigapp_v1.0.yaml
openapi: 3.0.3
info:
  title: Frig'App API
  description: |
    API REST de Frig'App — application mobile de gestion de stock de frigo collaboratif.
    
    ## Authentification
    Toutes les routes protégées nécessitent un token JWT passé dans le header `Authorization: Bearer <token>`.
    
    ## Temps réel
    Les modifications sur les items et les membres sont diffusées via WebSocket (STOMP over SockJS) sur `ws://api.frigapp.com/v1/ws`.
    
    ## Conventions
    - Toutes les réponses sont enveloppées dans `{ "status": "success"|"error", "data": ... }`
    - Les identifiants sont des UUID v4
    - Les timestamps sont en ISO 8601 UTC
  version: 1.0.0
  contact:
    name: Équipe Frig'App
servers:
  - url: https://api.frigapp.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Développement local (Spring Boot)
...
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── SCHEMAS ─────────────────────────────────────────────────────────────────
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
          example: success
        data:
          nullable: true
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
          example: error
        error:
          type: object
          properties:
            code:
              type: string
              example: FRIDGE_NOT_FOUND
            message:
              type: string
              example: Le frigo demandé n'existe pas.

    # ── USER ──────────────────────────────────────────────────────────────────
...
    FridgeRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Frigo maison
    FridgeSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440001
        name:
          type: string
          example: Frigo maison
        role:
          type: string
          enum: [owner, collaborator]
          example: owner
        itemCount:
          type: integer
          example: 12
        memberCount:
          type: integer
          example: 3
        urgentCount:
          type: integer
          description: Nombre de produits expirant dans les 3 prochains jours
          example: 2
        updatedAt:
          type: string
          format: date-time
          example: 2026-03-23T09:00:00Z
    FridgeFull:
      allOf:
        - $ref: '#/components/schemas/FridgeSummary'
        - type: object
          properties:
            ownerId:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time

    # ── MEMBER ────────────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Token JWT absent ou invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            error:
              code: UNAUTHORIZED
              message: Token JWT absent ou invalide.
    Forbidden:
      description: Action non autorisée pour ce rôle
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            error:
              code: FORBIDDEN
              message: Vous n'avez pas les droits pour effectuer cette action.
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadRequest:
      description: Données invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

# ── PATHS ─────────────────────────────────────────────────────────────────────
paths:
...
  /fridges:
    get:
      tags: [Fridges]
      summary: Lister mes frigos
      description: Retourne tous les frigos auxquels l'utilisateur appartient (possédés + collaborations).
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des frigos
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/FridgeSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Fridges]
      summary: Créer un frigo
      description: Crée un nouveau frigo. L'utilisateur courant devient automatiquement propriétaire.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FridgeRequest'
      responses:
        '201':
          description: Frigo créé
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FridgeFull'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /fridges/{fridgeId}:
    parameters:
      - name: fridgeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Identifiant du frigo

    get:
      tags: [Fridges]
      summary: Détail d'un frigo
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Détail du frigo
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FridgeFull'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Fridges]
      summary: Renommer un frigo
      description: Réservé au propriétaire du frigo.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FridgeRequest'
      responses:
        '200':
          description: Frigo renommé
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FridgeFull'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Fridges]
      summary: Supprimer un frigo
      description: Supprime le frigo et toutes ses données en cascade. Réservé au propriétaire.
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Frigo supprimé
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /fridges/{fridgeId}/leave:
    parameters:
      - name: fridgeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [Fridges]
      summary: Quitter un frigo
      description: Retire l'utilisateur courant des membres du frigo. Réservé aux collaborateurs.
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Frigo quitté
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Le propriétaire ne peut pas quitter son propre frigo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── MEMBERS ───────────────────────────────────────────────────────────────
...